Added a huge update
Added a combination puzzle that must be solved at the beginning of the game. I added jump scares and new sounds. Made the settings menu better by changing the layout and adding a variety of more options to choose from. Fixed a few bugs and reduced frame rate drops significantly